Add `filename` to exec/eval and `parse` to eval
These methods now accept a `filename` parameter, which will be used to
set the `filename` attribute on any resulting `SyntaxError` objects.

`eval` now accepts a `parse` parameter; if it is not True, then eval
will return a string containing the unparsed expression.

This also adds a bunch of tests for parsing, in preparation for moving
away from ast.literal_eval and building Python data structures directly
in a subsequent pull request.