Test mutiple calls to exec / handle resolve.Error
In https://github.com/caketop/pystarlark/pull/26 another change snuck
in to retain the globals across multiple calls to exec, so that it
could be used to define multiple things. While writing a test for this
behavor, I discovered that it is possible for starlark-go's Eval to
return a resolve.ErrorList, in addition to a starlark.EvalError or
a syntax.Error, so I had to add handling for that as well.