Build wheels for more platforms
Enable i686 and aarch64 builds for Linux, and universal2 builds for
macOS. Caught a couple portability problems in the process.

I made a mighty effort trying to get musl and Windows working
along the way but failed; musl is broken because of a bad interaction
with cgo and is probably beyond fixing short of patching the Go
compiler. Windows is suffering from what looks like garden-variety
linker errors and is probably salvagable, but I probably need someone
who knows more about building wheels on Windows to explain what I've
done wrong before I'll be able to fix it.